Adding recommendations function
The recommendation function is declared outside of route and is called in the route function in case the user hasn't chosen any filters and the query string is empty. If so, it is checked if the user has liked any restaurants - if the user_id is in the restaurant_user junction table, containing user ids and ids of the liked restaurants. If no restaurants have been liked, so the user_id is absent in the junction table, then all restaurants are fetched to the page according to descending order of their ratings. Else, up to 3 most frequent cuisines among the liked restaurants are identified and the most frequent values of dietary options and price range. Then, restaurants matching these criterias are found and fetched to the page. Once the user has applied any filters, restaurants based on them are fetched instead.