Adapting html for recommended restaurants data
The data from recommendations function has different length, as in case of recommended restaurants, coordinates are not required to be called from database. In this case it is important to specify in which row of data set the cuisine names are located, otherwise they won't be shown in the page.